+++ CHANGELOG.md
@@ -1,2 +1,8 @@
-- Fix power filters
-- Fix Consumables indicator
+[Full Changelog](https://github.com/enderneko/Cell/compare/r176-release...097786de90d25dcb21b8f30958ad9ab0369281c5)
+
+- Update for Augmentation
+- Add a "Fade out over time" option for textures
+- Add "Unit's Target" to spotlight frame
+- Remove CLEU auras from Raid Debuffs indicator
+- Update zhTW
+- [snippet] update IconFadeOut